refactor: use import type instead of reference
