+++ src/types/global.ts
@@ -1,0 +1,7 @@
+export {}
+declare global {
+  interface Element {
+    setAttribute(name: string, value: boolean): void
+    setAttribute(name: string, value: number): void
+  }
+}
