+++ src/types/linkDiv.ts
@@ -1,5 +1,5 @@
-type MainLineParams = { x1: number; y1: number; x2: number; y2: number }
-type SubLineParams = {
+export type MainLineParams = { x1: number; y1: number; x2: number; y2: number }
+export type SubLineParams = {
   pT: number
   pL: number
   pW: number
