--- types/mind-elixir/package.json
@@ -1,5 +1,0 @@
-{
-    "name": "mind-elixir",
-    "version": "1.0.0",
-    "typings": "index.d.ts"
-}
